@@ -13,12 +13,12 @@
 		}
 	],
 	"require": {
-		"php": "^5.3 || ^7.0",
+		"php": "^7.0",
 		"ext-sockets": ">=0",
 		"zendframework/zend-mail": "^2.3"
 	},
 	"require-dev": {
-		"phpunit/phpunit": "^4.8",
+		"phpunit/phpunit": "^6.2",
 		"phpmailer/phpmailer": "^5.2",
 		"phpstan/phpstan": "^0.7",
 		"thefox/utilities": "^1.1"
