@@ -2,9 +2,9 @@
 
 namespace TheFox\Test;
 
-use PHPUnit_Framework_TestCase;
+use PHPUnit\Framework\TestCase;
 
-class BasicTest extends PHPUnit_Framework_TestCase
+class BasicTest extends TestCase
 {
     public function testTest()
     {
