@@ -2,10 +2,10 @@
 
 namespace TheFox\Test;
 
-use PHPUnit_Framework_TestCase;
+use PHPUnit\Framework\TestCase;
 use TheFox\Smtp\Thread;
 
-class ThreadTest extends PHPUnit_Framework_TestCase
+class ThreadTest extends TestCase
 {
     public function testThread()
     {
