@@ -7,5 +7,5 @@ namespace TheFox\Smtp;
 class Smtpd
 {
     const NAME = 'SMTPd';
-    const VERSION = '0.4.0-dev.2';
+    const VERSION = '0.4.0-dev.4';
 }
