@@ -12,7 +12,7 @@ class StringParser
      * @var string
      */
     private $str = '';
-    
+
     private $len = 0;
 
     /**
@@ -37,10 +37,10 @@ class StringParser
 
     /**
      * StringParser constructor.
-     * @param $str
+     * @param string $str
      * @param null|int $argsMax
      */
-    public function __construct($str, $argsMax = null)
+    public function __construct(string $str, $argsMax = null)
     {
         $this->str = $str;
         $this->str = trim($this->str);
@@ -74,7 +74,7 @@ class StringParser
     /**
      * @param string $char
      */
-    private function charNew($char = '')
+    private function charNew(string $char = '')
     {
         if ($this->argsMax === null || $this->argsLen < $this->argsMax) {
             $this->fixPrev();
@@ -82,35 +82,31 @@ class StringParser
             $this->args[$this->argsId] = $char;
             $this->argsLen = count($this->args);
         }
-        /*else{
-            $this->charAppend($char);
-        }*/
     }
 
     /**
      * @param string $char
      */
-    private function charAppend($char)
+    private function charAppend(string $char)
     {
         if ($this->argsId != -1) {
             $this->args[$this->argsId] .= $char;
         }
-        /*else{
-            $this->charNew($char);
-        }*/
     }
 
-    public function parse()
+    /**
+     * @return array
+     */
+    public function parse(): array
     {
         $this->reset();
 
         $str = $this->str;
         $in = false;
-        $prevChar = ' ';
+        //$prevChar = ' ';
         $endChar = '';
 
         for ($pos = 0; $pos < $this->len; $pos++) {
-
             $char = $str[$pos];
             $nextChar = ($pos < $this->len - 1) ? $str[$pos + 1] : '';
 
@@ -151,7 +147,7 @@ class StringParser
                 }*/
             }
 
-            $prevChar = $char;
+            //$prevChar = $char;
         }
 
         $this->fixPrev();
